Correct phpstan errors

+++ modules/poker-game/app/PotHandler/PotHandler.php
@@ -26,7 +26,7 @@ class PotHandler
         $this->stackModel->change($amount, $playerId, $tableId);
     }
 
-    public function updatePot(int $betAmount, int $handId): void
+    public function updatePot(float $betAmount, int $handId): void
     {
         $pot = $this->potModel->find(['hand_id' => $handId]);
         $pot->update(['amount' => $pot->getAmount() + $betAmount]);
